+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 83.4 - 2025-06-05
+### Added
+- Dastardly Duos zone data
+
+### Fixed
+- Only temporarily obtainable achievements that are currently obtainable should have a green header color
+
 ## 83.3 - 2025-06-03
 ### Added
 - Flame Warden of Khaz Algar to the Midsummer category